% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhLoglikehood.r
\name{absValue}
\alias{absValue}
\title{Function to return absolute values}
\usage{
absValue(x)
}
\arguments{
\item{x}{Numeric, the input value.}
}
\value{
Numeric, the absolute value of the input.
}
\description{
Function to return absolute values
}
