% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputHelpers.R
\name{apply_burn_in}
\alias{apply_burn_in}
\title{Apply Burn-In}
\usage{
apply_burn_in(results, burn_in)
}
\arguments{
\item{results}{A list of MCMC chain results.}

\item{burn_in}{The fraction or proportion of results to discard as burn-in (0 to 1). The default is no burn-in, burn_in=0.}
}
\value{
A list of results with burn-in applied.
}
\description{
Apply Burn-In
}
