% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputHelpers.R
\name{apply_thinning}
\alias{apply_thinning}
\title{Apply Thinning}
\usage{
apply_thinning(results, thinning_factor)
}
\arguments{
\item{results}{A list of MCMC chain results.}

\item{thinning_factor}{The factor by which to thin the results (positive integer). The default thinning factor is 1, which implies no thinning.}
}
\value{
A list of results with thinning applied.
}
\description{
Apply Thinning
}
