% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/docData.R
\docType{data}
\name{baseline_data_default}
\alias{baseline_data_default}
\title{Default Baseline Data}
\format{
A data frame with 94 rows and 3 variables:
\describe{
\item{Age}{Age in years}
\item{Female}{Female baseline risk}
\item{Male}{Male baseline risk}
}

A data frame with 94 rows and 3 variables:
\describe{
\item{Age}{Age in years (1 to 94)}
\item{Female}{Penetrance rate for females}
\item{Male}{Penetrance rate for males}
}
}
\usage{
baseline_data_default

baseline_data_default
}
\description{
This dataset contains age-specific cancer penetrance rates for both females and males.
As an example, the data is derived from the SEER program for colorectal cancer females and males.
}
\keyword{datasets}
