% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeAges.R
\name{calcPedDegree}
\alias{calcPedDegree}
\title{Calculate Degree of Relationship in Pedigree}
\usage{
calcPedDegree(data)
}
\arguments{
\item{data}{A data frame containing the data with columns 'individual', 'father', 'mother', 'sex', 'aff', 'family', and 'isProband'.}
}
\value{
The data frame with an additional column 'degree_of_relationship' indicating the degree of relationship for each individual.
}
\description{
This function calculates the degree of relationship for individuals in a dataset based on their pedigree.
}
