% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{calculate_lifetime_risk}
\alias{calculate_lifetime_risk}
\title{Calculate Lifetime Risk of Cancer}
\usage{
calculate_lifetime_risk(cancer_type, gene, race, sex, type, db)
}
\arguments{
\item{cancer_type}{The type of cancer for which the risk is being calculated.}

\item{gene}{The gene of interest for which the risk is being calculated.}

\item{race}{The race of the individual.}

\item{sex}{The sex of the individual, Female or Male.}

\item{type}{The type of penetrance calculation, Net or Crude.}

\item{db}{The dataset used for the calculation.}
}
\value{
A list containing the cumulative risk (\code{cumulative_risk}) and the total
probability (\code{total_prob}) of developing the specified cancer.
}
\description{
This function calculates the lifetime risk of a specific type of cancer based
on genetic and demographic factors using data in the format of the PanelPRO Database.
}
