% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhLoglikehood.r
\name{calculateBaseline}
\alias{calculateBaseline}
\title{Calculate Baseline Risk}
\usage{
calculateBaseline(cancer_type, gene, race, type, db)
}
\arguments{
\item{cancer_type}{The type of cancer for which the risk is being calculated.}

\item{gene}{The gene of interest for which the risk is being calculated.}

\item{race}{The race of the individual.}

\item{type}{The type of penetrance calculation.}

\item{db}{The dataset used for the calculation, containing penetrance data.}
}
\value{
A matrix of penetrance data for the specified parameters.
}
\description{
This function extracts the penetrance data for a specified cancer type, gene,
race, and penetrance type from the provided database.
}
