% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputHelpers.R
\name{combine_chains_noSex}
\alias{combine_chains_noSex}
\title{Combine Chains for Non-Sex-Specific Estimation}
\usage{
combine_chains_noSex(results)
}
\arguments{
\item{results}{A list of MCMC chain results, where each element contains posterior samples of parameters.}
}
\value{
A list with combined results, including samples for median, threshold, first quartile, asymptote values,
log-likelihoods, and log-acceptance ratios.
}
\description{
Combines the posterior samples from multiple MCMC chains for non-sex-specific estimations.
}
