% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/priorElicitation.R
\docType{data}
\name{distribution_data_default}
\alias{distribution_data_default}
\title{Default Distribution Data}
\format{
A data frame with the following columns:
\describe{
\item{age}{Age values (NA for default).}
\item{at_risk}{Proportion of people at risk (NA for default).}
}
}
\usage{
distribution_data_default
}
\description{
Default data frame structure with row names for use in the \code{makePriors} function.
}
\keyword{datasets}
