% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeAges.R
\name{drawEmpirical}
\alias{drawEmpirical}
\title{Draw Ages Using the Inverse CDF Method from Empirical Density}
\usage{
drawEmpirical(empirical_density, sex = NA)
}
\arguments{
\item{empirical_density}{A density object or a list of density objects (if sex-specific) containing the empirical density of ages.}

\item{sex}{Numeric, the sex of the individual (1 for male, 2 for female). Required if \code{empirical_density} is a list.}
}
\value{
A single age value drawn from the empirical density data.
}
\description{
This function draws ages using the inverse CDF method from empirical density data.
}
