% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputHelpers.R
\name{generate_density_plots}
\alias{generate_density_plots}
\title{Generate Posterior Density Plots}
\usage{
generate_density_plots(data)
}
\arguments{
\item{data}{A list with combined results.}
}
\description{
Generates histograms of the posterior samples for the different parameters
}
