% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputHelpers.R
\name{generate_summary}
\alias{generate_summary}
\title{Generate Summary}
\usage{
generate_summary(data)
}
\arguments{
\item{data}{A list with combined results.}
}
\value{
A summary data frame containing Median, threshold, First Quartile, and Asymptote Value.
}
\description{
Function to generate summary statistics
}
