% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noSex.R, R/outputHelpers.R
\name{generate_summary_noSex}
\alias{generate_summary_noSex}
\title{Generate Summary}
\usage{
generate_summary_noSex(data)

generate_summary_noSex(data)
}
\arguments{
\item{data}{A list containing combined results of MCMC chains, typically the output of \code{combine_chains_noSex}.}
}
\value{
A summary data frame containing Median, threshold, First Quartile, and Asymptote Value.

A summary data frame containing median, threshold, first quartile, and asymptote values.
}
\description{
Function to generate summary statistics

Generates summary statistics for the combined MCMC results for non-sex-specific estimations.
}
