% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeAges.R
\name{imputeAgesInit}
\alias{imputeAgesInit}
\title{Initialize Ages Using a Uniform Distribution}
\usage{
imputeAgesInit(data, threshold, max_age)
}
\arguments{
\item{data}{A data frame containing the data.}

\item{threshold}{The minimum age value for the uniform distribution.}

\item{max_age}{The maximum age value for the uniform distribution.}
}
\value{
A list containing the updated data frame and the indices of the rows where ages were initialized.
}
\description{
This function initializes ages for individuals in a dataset using a uniform distribution.
}
