% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docData.R
\docType{data}
\name{out_sim}
\alias{out_sim}
\title{Simulated Output Data}
\format{
A list with the following components:
\describe{
\item{summary_stats}{A data frame with 18000 observations of 8 variables:
\describe{
\item{Median_Male}{numeric, Median value for males}
\item{Median_Female}{numeric, Median value for females}
\item{Threshold_Male}{numeric, Threshold value for males}
\item{Threshold_Female}{numeric, Threshold value for females}
\item{First_Quartile_Male}{numeric, First quartile value for males}
\item{First_Quartile_Female}{numeric, First quartile value for females}
\item{Asymptote_Male}{numeric, Asymptote value for males}
\item{Asymptote_Female}{numeric, Asymptote value for females}
}
}
\item{density_plots}{A list of 1 element, mfrow: integer vector of length 2}
\item{trace_plots}{A list of 1 element, mfrow: integer vector of length 2}
\item{penetrance_plot}{A list of 2 elements: rect and text}
\item{pdf_plots}{A list of 2 elements: rect and text}
\item{combined_chains}{A list of 19 numeric vectors with 18000 elements each}
\item{results}{A list of 1 element which is a list of 24 elements, each with 18000 elements}
\item{data}{A data frame with 4727 observations of 9 variables:
\describe{
\item{individual}{integer, Individual ID}
\item{isProband}{numeric, Indicator if the individual is a proband}
\item{family}{integer, Family ID}
\item{mother}{numeric, Mother's ID}
\item{father}{numeric, Father's ID}
\item{aff}{numeric, Affected status}
\item{sex}{numeric, Sex of the individual}
\item{age}{numeric, Age of the individual}
\item{geno}{character, Genotype}
}
}
}
}
\usage{
data(out_sim)
}
\description{
This dataset contains the simulated output data for the penetrance package.
}
\examples{
data(out_sim)
head(out_sim$summary_stats)
}
\keyword{datasets}
