% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PenEstimMain.R
\name{PenEstim}
\alias{PenEstim}
\title{Bayesian Inference using Independent Metropolis-Hastings for Penetrance Estimation}
\usage{
PenEstim(
  pedigree,
  twins,
  n_chains = 1,
  n_iter_per_chain = 10000,
  ncores = 6,
  max_age = 94,
  baseline_data = baseline_data_default,
  removeProband = FALSE,
  ageImputation = FALSE,
  median_max = TRUE,
  BaselineNC = TRUE,
  var = c(0.1, 0.1, 2, 2, 5, 5, 5, 5),
  burn_in = 0,
  thinning_factor = 1,
  distribution_data = distribution_data_default,
  af = 1e-04,
  max_penetrance = 1,
  sample_size = NULL,
  ratio = NULL,
  prior_params = prior_params_default,
  risk_proportion = risk_proportion_default,
  summary_stats = TRUE,
  rejection_rates = TRUE,
  density_plots = TRUE,
  penetrance_plot = TRUE,
  probCI = 0.95
)
}
\arguments{
\item{pedigree}{A data frame containing the pedigree data in the required format. It should include the following columns:
\itemize{
\item \code{PedigreeID}:  A numeric value representing the unique identifier for each family There should be no duplicated entries.
\item \code{ID}: A numeric value representing the unique identifier for each individual. There should be no duplicated entries.
\item \code{Sex}: A numeric value where \code{0} indicates female and \code{1} indicates male. Missing entries are not currently supported.
\item \code{MotherID}: A numeric value representing the unique identifier for an individual's mother.
\item \code{FatherID}: A numeric value representing the unique identifier for an individual's father.
\item \code{isProband}: A numeric value where \code{1} indicates the individual is a proband and \code{0} otherwise.
\item \code{CurAge}: A numeric value indicating the age of censoring (current age if the person is alive or age at death if the person is deceased). Allowed ages range from \code{1} to \code{94}.
\item \code{isAff}: A numeric value indicating the affection status of cancer, with \code{1} for diagnosed individuals and \code{0} otherwise. Missing entries are not supported.
\item \code{Age}: A numeric value indicating the age of cancer diagnosis, encoded as \code{NA} if the individual was not diagnosed. Allowed ages range from \code{1} to \code{94}.
\item \code{Geno}: A column for germline testing or tumor marker testing results. Positive results should be coded as \code{1}, negative results as \code{0}, and unknown results as \code{NA} or left empty.
}}

\item{twins}{Identical twins or triplets in the family can be specifed. For example, to indicate that \code{ora024} and \code{ora027} are identical twins, and so are \code{aey063} and \code{aey064},
then we can use the following as the twins arguement: twins <- list(c("ora024", "ora027"), c("aey063", "aey064"))}

\item{n_chains}{Number of chains for parallel computation. Default is 1.}

\item{n_iter_per_chain}{Number of iterations for each chain. Default is 10000.}

\item{ncores}{Number of cores for parallel computation. Default is 6.}

\item{max_age}{Maximum age considered for analysis. Default is 94.}

\item{baseline_data}{Data for the baseline risk estimates (probability of developing cancer), such as population-level risk from a cancer registry. Default is the allele frequency for MLH1 from the PanelPRO database.}

\item{removeProband}{Logical, indicating whether to remove probands from the analysis. Default is FALSE.}

\item{ageImputation}{Logical, indicating whether to perform age imputation. Default is FALSE.}

\item{median_max}{Boolean indicating whether to use the baseline median age or max_age as an upper bound for the median proposal. Default is TRUE.}

\item{BaselineNC}{Boolean indicating that the non-carrier penetrance is assumed to be the baseline penetrance. Default is TRUE.}

\item{var}{Vector of variances for the proposal distribution. Default is c(0.1, 0.1, 2, 2, 5, 5, 5, 5).}

\item{burn_in}{Fraction of results to discard as burn-in (0 to 1). Default is 0 (no burn-in).}

\item{thinning_factor}{Factor by which to thin the results. Default is 1 (no thinning).}

\item{distribution_data}{Data for generating prior distributions.}

\item{af}{Allele frequency for the risk allele. Default is 0.0001.}

\item{max_penetrance}{Maximum penetrance considered for analysis. Default is 1.}

\item{sample_size}{Optional sample size for distribution generation.}

\item{ratio}{Optional ratio parameter for distribution generation.}

\item{prior_params}{Parameters for prior distributions.}

\item{risk_proportion}{Proportion of risk for distribution generation.}

\item{summary_stats}{Boolean indicating whether to include summary statistics in the output. Default is TRUE.}

\item{rejection_rates}{Boolean indicating whether to include rejection rates in the output. Default is TRUE.}

\item{density_plots}{Boolean indicating whether to include density plots in the output. Default is TRUE.}

\item{penetrance_plot}{Boolean indicating whether to include penetrance plots in the output. Default is TRUE.}

\item{probCI}{Probability level for confidence intervals in penetrance plots. Default is 0.95.}
}
\value{
A list containing combined results from all chains, along with optional statistics and plots.
}
\description{
This function implements the Independent Metropolis-Hastings algorithm for Bayesian
penetrance estimation of cancer risk. It utilizes parallel computing to run multiple
chains and provides various options for analyzing and visualizing the results.
}
