% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputHelpers.R
\name{plot_acf}
\alias{plot_acf}
\title{Plot Autocorrelation for Multiple MCMC Chains (Posterior Samples)}
\usage{
plot_acf(results, n_chains, max_lag = 50)
}
\arguments{
\item{results}{A list of MCMC chain results.}

\item{n_chains}{The number of chains.}

\item{max_lag}{Integer, the maximum lag to be considered for the autocorrelation plot. Default is 50.}
}
\value{
A series of autocorrelation plots for each chain.
}
\description{
This function plots the autocorrelation for sex-specific or non-sex-specific posterior samples across multiple MCMC chains.
It defaults to key parameters like \code{asymptote_male_samples}, \code{asymptote_female_samples}, etc.
}
