% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputHelpers.R
\name{plot_loglikelihood}
\alias{plot_loglikelihood}
\title{Plot Log-Likelihood for Multiple MCMC Chains}
\usage{
plot_loglikelihood(results, n_chains)
}
\arguments{
\item{results}{A list of MCMC chain results, each containing the \code{loglikelihood_current} values.}

\item{n_chains}{The number of chains.}
}
\value{
A series of log-likelihood plots for each chain.
}
\description{
This function plots the log-likelihood values across iterations for multiple MCMC chains.
It helps visualize the convergence of the chains based on the log-likelihood values.
}
