% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputHelpers.R
\name{printRejectionRates}
\alias{printRejectionRates}
\title{Print Rejection Rates}
\usage{
printRejectionRates(results)
}
\arguments{
\item{results}{A list of MCMC chain results.}
}
\description{
Print Rejection Rates
}
