% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/priorElicitation.R
\docType{data}
\name{prior_params_default}
\alias{prior_params_default}
\title{Default Prior Parameters}
\format{
A list of prior distribution parameters

A list with the following components:
\describe{
\item{asymptote}{A list with components \code{g1} and \code{g2}, default values for the asymptote parameters.}
\item{threshold}{A list with components \code{min} and \code{max}, default values for the threshold parameters.}
\item{median}{A list with components \code{m1} and \code{m2}, default values for the median parameters.}
\item{first_quartile}{A list with components \code{q1} and \code{q2}, default values for the first quartile parameters.}
}
}
\usage{
prior_params_default

prior_params_default
}
\description{
Default parameters for the prior distributions used in the \code{makePriors} function.
}
\keyword{datasets}
