% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/priorElicitation.R
\docType{data}
\name{risk_proportion_default}
\alias{risk_proportion_default}
\title{Default Risk Proportions}
\format{
A data frame of risk proportions

A data frame with the following columns:
\describe{
\item{median}{Proportion of people at risk at the median age.}
\item{first_quartile}{Proportion of people at risk at the first quartile age.}
\item{max_age}{Proportion of people at risk at the maximum age.}
}
}
\usage{
risk_proportion_default

risk_proportion_default
}
\description{
Default proportions of people at risk used in the \code{makePriors} function.
}
\keyword{datasets}
