% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeAges.R
\name{calculateEmpiricalDensity}
\alias{calculateEmpiricalDensity}
\title{Calculate Empirical Density for Non-Affected Individuals}
\usage{
calculateEmpiricalDensity(
  data,
  aff_column = "aff",
  age_column = "age",
  n_points = 10000
)
}
\arguments{
\item{data}{A data frame containing the data.}

\item{aff_column}{The name of the column indicating affection status (default is "aff").}

\item{age_column}{The name of the column indicating ages (default is "age").}

\item{n_points}{The number of points to use in the density estimation (default is 10000).}
}
\value{
A density object representing the empirical density of ages.
}
\description{
This function calculates the empirical density for ages of non-affected individuals in a dataset.
}
