% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeAges.R
\name{calculateEmpiricalDensity}
\alias{calculateEmpiricalDensity}
\title{Calculate Empirical Density for Non-Affected Individuals}
\usage{
calculateEmpiricalDensity(
  data,
  aff_column = "aff",
  age_column = "age",
  sex_column = "sex",
  geno_column = "geno",
  n_points = 10000,
  sex_specific = TRUE
)
}
\arguments{
\item{data}{A data frame containing the data.}

\item{aff_column}{Character, the name of the column indicating affection status. Default is "aff".}

\item{age_column}{Character, the name of the column indicating ages. Default is "age".}

\item{sex_column}{Character, the name of the column indicating sex. Default is "sex".}

\item{geno_column}{Character, the name of the column indicating genotype. Default is "geno".}

\item{n_points}{Integer, the number of points to use in the density estimation. Default is 10000.}
}
\value{
A list of density objects representing the empirical density of ages for different groups.
}
\description{
This function calculates the empirical density for ages of non-affected individuals in a dataset,
differentiating by sex and whether the individual was tested (has a non-NA 'geno' value).
}
