% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeAges.R
\name{drawEmpirical}
\alias{drawEmpirical}
\title{Draw Ages Using the Inverse CDF Method from Empirical Density}
\usage{
drawEmpirical(empirical_density)
}
\arguments{
\item{empirical_density}{A density object containing the empirical density of ages.}
}
\value{
A single age value drawn from the empirical density data.
}
\description{
This function draws ages using the inverse CDF method from empirical density data.
}
