% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeAges.R
\name{drawEmpirical}
\alias{drawEmpirical}
\title{Draw Ages Using the Inverse CDF Method from Empirical Density}
\usage{
drawEmpirical(empirical_density, sex, tested, sex_specific = TRUE)
}
\arguments{
\item{empirical_density}{A list of density objects containing the empirical density of ages for different groups.}

\item{sex}{Numeric, the sex of the individual (1 for male, 2 for female).}

\item{tested}{Logical, indicating whether the individual was tested (has a non-NA 'geno' value).}

\item{sex_specific}{Logical, indicating whether the imputation should be sex-specific. Default is TRUE.}
}
\value{
A single age value drawn from the appropriate empirical density data.
}
\description{
This function draws ages using the inverse CDF method from empirical density data,
based on sex and whether the individual was tested.
}
