% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputHelpers.R
\name{generate_summary}
\alias{generate_summary}
\title{Generate Summary}
\usage{
generate_summary(data, verbose = FALSE)
}
\arguments{
\item{data}{A list with combined results.}

\item{verbose}{Logical, whether to print summary to console. Default is FALSE.}
}
\value{
A data.frame containing summary statistics (min, 1st quartile, median, mean, 3rd quartile, max)
for each parameter.
}
\description{
Function to generate summary statistics
}
