% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputHelpers.R
\name{generate_summary_noSex}
\alias{generate_summary_noSex}
\title{Generate Summary for Non-Sex-Specific Estimation}
\usage{
generate_summary_noSex(data, verbose = FALSE)
}
\arguments{
\item{data}{A list containing combined results of MCMC chains, typically the output of \code{combine_chains_noSex}.}

\item{verbose}{Logical, whether to print summary to console. Default is FALSE.}
}
\value{
A data.frame containing summary statistics (min, 1st quartile, median, mean, 3rd quartile, max)
for median, threshold, first quartile, and asymptote values.
}
\description{
Generates summary statistics for the combined MCMC results for non-sex-specific estimations.
}
