% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeAges.R
\name{imputeAges}
\alias{imputeAges}
\title{Impute Ages Based on Affection Status and Sex}
\usage{
imputeAges(
  data,
  na_indices,
  baseline_male,
  baseline_female,
  alpha_male,
  beta_male,
  delta_male,
  alpha_female,
  beta_female,
  delta_female,
  empirical_density,
  max_age
)
}
\arguments{
\item{data}{A data frame containing the data.}

\item{na_indices}{A vector of indices indicating the rows in the data where ages need to be imputed.}

\item{baseline_male}{A data frame containing baseline data for males.}

\item{baseline_female}{A data frame containing baseline data for females.}

\item{alpha_male}{Shape parameter for the Weibull distribution for males.}

\item{beta_male}{Scale parameter for the Weibull distribution for males.}

\item{delta_male}{Location parameter for the Weibull distribution for males.}

\item{alpha_female}{Shape parameter for the Weibull distribution for females.}

\item{beta_female}{Scale parameter for the Weibull distribution for females.}

\item{delta_female}{Location parameter for the Weibull distribution for females.}

\item{empirical_density}{A density object containing the empirical density of ages.}

\item{max_age}{Integer, the maximum age considered in the analysis.}
}
\value{
The data frame with imputed ages.
}
\description{
This function imputes ages for individuals in a dataset based on their affection
status and sex using either Weibull, baseline, or empirical distribution.
}
