% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imputeAges.R
\name{imputeAges}
\alias{imputeAges}
\title{Impute Ages Based on Affection Status and Sex}
\usage{
imputeAges(
  data,
  na_indices,
  baseline_male = NULL,
  baseline_female = NULL,
  alpha_male = NULL,
  beta_male = NULL,
  delta_male = NULL,
  alpha_female = NULL,
  beta_female = NULL,
  delta_female = NULL,
  baseline = NULL,
  alpha = NULL,
  beta = NULL,
  delta = NULL,
  max_age,
  sex_specific = TRUE,
  max_attempts = 100,
  geno_freq,
  trans,
  lik
)
}
\arguments{
\item{data}{A data frame containing the individual data, including columns for age, sex, and affection status.}

\item{na_indices}{A vector of indices indicating the rows in the data where ages need to be imputed.}

\item{baseline_male}{A data frame containing baseline data for males, with columns 'cum_prob' and 'age'.}

\item{baseline_female}{A data frame containing baseline data for females, with columns 'cum_prob' and 'age'.}

\item{alpha_male}{Numeric, shape parameter for the Weibull distribution for males.}

\item{beta_male}{Numeric, scale parameter for the Weibull distribution for males.}

\item{delta_male}{Numeric, location parameter for the Weibull distribution for males.}

\item{alpha_female}{Numeric, shape parameter for the Weibull distribution for females.}

\item{beta_female}{Numeric, scale parameter for the Weibull distribution for females.}

\item{delta_female}{Numeric, location parameter for the Weibull distribution for females.}

\item{baseline}{A data frame containing baseline data (used for non-sex-specific analysis) with columns 'cum_prob' and 'age'.}

\item{alpha}{Numeric, shape parameter for the Weibull distribution (used for non-sex-specific analysis).}

\item{beta}{Numeric, scale parameter for the Weibull distribution (used for non-sex-specific analysis).}

\item{delta}{Numeric, location parameter for the Weibull distribution (used for non-sex-specific analysis).}

\item{max_age}{Integer, the maximum age considered in the analysis.}

\item{sex_specific}{Logical, indicating whether the imputation should be sex-specific. Default is TRUE.}

\item{max_attempts}{Integer, the maximum number of attempts to get a valid age. Default is 100.}

\item{geno_freq}{A numeric vector representing genotype frequencies.}

\item{trans}{A matrix or data frame representing transmission probabilities.}

\item{lik}{A likelihood matrix returned by the mhloglikelihood_clipp function, used in genotype probability calculations.}
}
\value{
The data frame with imputed ages.
}
\description{
This function imputes ages for individuals in a dataset based on their affection
status and sex using either Weibull, baseline, or empirical distribution.
}
