% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhLoglikehood.r
\name{mhLogLikelihood_clipp}
\alias{mhLogLikelihood_clipp}
\title{Calculate Log Likelihood using clipp Package}
\usage{
mhLogLikelihood_clipp(
  paras,
  families,
  twins,
  max_age,
  baseline_data,
  af,
  geno_freq,
  trans,
  BaselineNC,
  ncores
)
}
\arguments{
\item{paras}{Numeric vector, the parameters for the Weibull distribution and scaling factors.
Should contain in order: gamma_male, gamma_female, delta_male, delta_female,
given_median_male, given_median_female, given_first_quartile_male, given_first_quartile_female.}

\item{families}{Data frame, containing pedigree information with columns for 'sex', 'age', 'aff' (affection status), and 'geno' (genotype).}

\item{twins}{Information on monozygous twins or triplets in the pedigrees.}

\item{max_age}{Integer, maximum age considered in the analysis.}

\item{baseline_data}{Numeric matrix, baseline risk data for each age by sex. Rows correspond to sex (1 for male, 2 for female) and columns to age.}

\item{af}{Numeric, allele frequency of the risk allele in the population.}

\item{BaselineNC}{Logical, indicates if non-carrier penetrance should be based on the baseline data or the calculated non-carrier penetrance.}

\item{ncores}{Integer, number of cores to use for parallel computation.}
}
\value{
Numeric, the calculated log likelihood.
}
\description{
This function calculates the log likelihood using the clipp package for a set of parameters and data.
}
\references{
Details about the clipp package and methods can be found in the package documentation.
}
