% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputHelpers.R
\name{plot_penetrance}
\alias{plot_penetrance}
\title{Plot Weibull Distribution with Credible Intervals}
\usage{
plot_penetrance(combined_chains, prob, max_age, sex = "NA")
}
\arguments{
\item{combined_chains}{List of combined MCMC chain results containing posterior samples
for penetrance parameters.}

\item{prob}{Numeric, probability level for confidence intervals (between 0 and 1).}

\item{max_age}{Integer, maximum age to plot.}

\item{sex}{Character, one of "Male", "Female", or "NA" for non-sex-specific. Default is "NA".}
}
\value{
A plot showing the Weibull distribution with credible intervals.
}
\description{
This function plots the Weibull distribution with credible intervals for the given MCMC results.
It allows for visualization of penetrance curves based on the posterior samples.
}
