% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputHelpers.R
\name{plot_trace}
\alias{plot_trace}
\title{Plot Trace}
\usage{
plot_trace(results, n_chains)
}
\arguments{
\item{results}{A list of MCMC chain results.}

\item{n_chains}{The number of chains.}
}
\description{
Plot Trace
}
