% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{transformDF}
\alias{transformDF}
\title{Transform Data Frame}
\usage{
transformDF(df)
}
\arguments{
\item{df}{The input data frame in the usual PanelPRO format.}
}
\value{
The transformed data frame in the format required for clipp.
}
\description{
This function transforms a data frame from the standard format used in PanelPRO
into the required format which conforms to the requirements of penetrance (and clipp).
}
